Add support for running changelog generation from a task and implement it in GradleUtils.
