+++ build.gradle
@@ -1,4 +1,5 @@
 import net.minecraftforge.gradleutils.GradleUtils
+import net.minecraftforge.gradleutils.tasks.GenerateChangelogTask
 
 plugins {
     id 'java-gradle-plugin'
@@ -28,6 +29,7 @@ task copyGradleUtils(type: Copy) {
     into 'src/main/groovy'
     from('buildSrc/src/main/groovy') {
         include '**/*.groovy'
+        include '**/*.java'
     }
 }
 
@@ -105,3 +107,10 @@ publishing {
         maven GradleUtils.getPublishingForgeMaven(project)
     }
 }
+
+task generateChangelog(type: GenerateChangelogTask) {
+    projectUrl = GradleUtils.buildProjectUrl("GradleUtils")
+    usesMarkdown = false
+}
+
+build.dependsOn generateChangelog
